Example Test Plans/Example Test Plans/Student 4/GitLab_Tests_4/UserInterface.java
public class UserInterface {
    public String selectUserType(String userType) {
        // Implementation to simulate user type selection
        if (userType.equals("1")) {
            return "Community";
        } else if (userType.equals("2")) {
            return "Encost";
        } else {
            return "Invalid";
        }
    }

    public boolean authenticateEncostUser(String username, String password) {
        // Implementation to authenticate Encost User
        return username.equals("encost_user") && password.equals("password");
    }

    public boolean loadCustomDataset(String username, String datasetFileName) {
        // Example: Validate user privileges and dataset file format
        return username.equals("encost_user") && datasetFileName.endsWith(".csv");
    }

    public boolean visualizeGraph(String userType) {
        // Implementation to visualize a graph based on user type
        return userType.equals("1") || userType.equals("2");
    }

    public String viewSummaryStatistics(String userType) {
        // Implementation to view summary statistics based on user type
        if (userType.equals("encost_user")) {
            return "Summary statistics for Encost User...";
        } else {
            return "Summary statistics for Community User...";
        }
    }
}
